refactor: split template loading out of company_helpers to clear the module-size budget
