+++ src/synthorg/api/controllers/setup/company_helpers.py
@@ -1,24 +1,23 @@
-"""Company-side setup helpers: locale handling, template loading, password length.
+"""Company-side setup helpers: locale handling, persistence, password length.
 
-Covers company metadata, template resolution, locale storage, and
-setup-complete checks. Agent-side helpers (bootstrap, model
-selection, tier coverage) live in ``setup._status_checks``,
-``setup._runtime_wiring``, and ``setup._embedder_setup``.
+Covers company metadata + persistence, locale storage, and setup-complete
+checks. Template loading/resolution lives in ``setup._template_helpers``;
+agent-side helpers (bootstrap, model selection, tier coverage) live in
+``setup._status_checks``, ``setup._runtime_wiring``, and
+``setup._embedder_setup``.
 """
 
 import json
 from collections.abc import Sequence
 from typing import Final, NamedTuple
 
-from synthorg.api.controllers.setup_agents import departments_to_json
 from synthorg.core.auth.config import AuthConfig
 from synthorg.core.collections import dedupe_preserving_order
 from synthorg.core.domain_errors import (
     ConflictError,
-    NotFoundError,
     ValidationError,
 )
-from synthorg.observability import get_logger, safe_error_description
+from synthorg.observability import get_logger
 from synthorg.observability.events.setup import (
     SETUP_ALREADY_COMPLETE,
     SETUP_COMPLETE_CHECK_ERROR,
@@ -26,14 +25,10 @@ from synthorg.observability.events.setup import (
     SETUP_NAME_LOCALES_INVALID,
     SETUP_STATUS_SETTINGS_DEFAULT_USED,
     SETUP_STATUS_SETTINGS_UNAVAILABLE,
-    SETUP_TEMPLATE_INVALID,
-    SETUP_TEMPLATE_NOT_FOUND,
 )
 from synthorg.settings.enums import SettingSource
 from synthorg.settings.errors import SettingNotFoundError
 from synthorg.settings.service_protocol import SettingsServiceProtocol
-from synthorg.templates.loader import LoadedTemplate
-from synthorg.templates.schema import CompanyTemplate
 
 logger = get_logger(__name__)
 
@@ -351,40 +346,6 @@ async def check_setup_not_complete(
         raise ConflictError(msg)
 
 
-class TemplateResult(NamedTuple):
-    """Result of template resolution."""
-
-    departments_json: str
-    department_count: int
-    template_applied: str | None
-    template: CompanyTemplate | None
-    loaded: LoadedTemplate | None
-
-
-def resolve_template(template_name: str | None) -> TemplateResult:
-    """Validate template and extract department data.
-
-    Returns:
-        ``TemplateResult`` instance. ``loaded`` carries the full
-        :class:`LoadedTemplate` so callers can render through the shared
-        renderer pipeline (resolving inheritance) rather than the raw template.
-    """
-    if template_name is None:
-        return TemplateResult("", 0, None, None, None)
-
-    loaded = load_template_safe(template_name)
-    departments_json = departments_to_json(
-        loaded.template.departments,
-    )
-    return TemplateResult(
-        departments_json,
-        len(loaded.template.departments),
-        template_name,
-        loaded.template,
-        loaded,
-    )
-
-
 class CompanyPersist(NamedTuple):
     """The backend-owned company-level fields written at company creation.
 
@@ -428,45 +389,3 @@ async def persist_company_settings(
     if fields.budget is not None:
         await settings_svc.set("company", "budget", str(fields.budget))
     await settings_svc.set("company", "company_name", fields.company_name)
-
-
-def load_template_safe(template_name: str) -> LoadedTemplate:
-    """Load a template by name with API-friendly error handling.
-
-    Args:
-        template_name: Template name to load.
-
-    Returns:
-        ``LoadedTemplate`` instance.
-
-    Raises:
-        NotFoundError: If the template does not exist.
-        ValidationError: If it fails to render or validate.
-    """
-    from synthorg.templates.errors import (  # noqa: PLC0415
-        TemplateNotFoundError,
-        TemplateRenderError,
-        TemplateValidationError,
-    )
-    from synthorg.templates.loader import (  # noqa: PLC0415
-        load_template,
-    )
-
-    try:
-        return load_template(template_name)
-    except TemplateNotFoundError as exc:
-        msg = f"Template {template_name!r} not found"
-        logger.warning(
-            SETUP_TEMPLATE_NOT_FOUND,
-            template=template_name,
-        )
-        raise NotFoundError(msg) from exc
-    except (TemplateRenderError, TemplateValidationError) as exc:
-        msg = f"Template {template_name!r} is invalid: {safe_error_description(exc)}"
-        logger.warning(
-            SETUP_TEMPLATE_INVALID,
-            template=template_name,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-        raise ValidationError(msg) from exc
